chore #144: 디렉토리 리펙토링
