+++ Project_Timer/Data/Response/TTErrorResponse.swift
@@ -1,5 +1,5 @@
 //
-//  ErrorResponse.swift
+//  TTErrorResponse.swift
 //  Project_Timer
 //
 //  Created by Kang Minsang on 2024/05/15.
@@ -9,7 +9,7 @@
 import Foundation
 
 /// TiTi 에러
-struct ErrorResponse: Decodable {
+struct TTErrorResponse: Decodable {
     let code: String
     let message: String
     let errors: [String: String]
