+++ Project_Timer/Data/Response/TTResponse.swift
@@ -1,0 +1,14 @@
+//
+//  TTResponse.swift
+//  Project_Timer
+//
+//  Created by Kang Minsang on 2024/06/29.
+//  Copyright © 2024 FDEE. All rights reserved.
+//
+
+import Foundation
+
+struct TTResponse: Decodable {
+    let code: String
+    let message: String
+}
