+++ Project_Timer/Data/lagacy/NetworkController.swift
@@ -7,6 +7,23 @@
 //
 
 import Foundation
+import Alamofire
+
+protocol TiTiFunctionsFetchable {
+    func getTiTiFunctions_lagacy(completion: @escaping (Result<[FunctionInfo], NetworkError>) -> Void)
+}
+
+protocol UpdateHistoryFetchable {
+    func getUpdateHistorys_lagacy(completion: @escaping (Result<[UpdateHistoryInfo], NetworkError>) -> Void)
+}
+
+protocol YoutubeLinkFetchable {
+    func getYoutubeLink_lagacy(completion: @escaping (Result<YoutubeLinkResponse, NetworkError>) -> Void)
+}
+
+protocol SurveysFetchable {
+    func getSurveys_lagacy(completion: @escaping (Result<[SurveyInfo], NetworkError>) -> Void)
+}
 
 final class NetworkController {
     let network: Network
